Improve popup contrast styles.
